WIP: better file structure
