+++ front/app/auth/controllers.py
@@ -1,31 +1,15 @@
 import functools
-
 from flask import (
     Blueprint, flash, g, redirect, render_template, request, session, url_for
 )
-from flask_wtf import FlaskForm
-from wtforms import StringField, PasswordField, SubmitField
-from wtforms.validators import InputRequired, Email, Length
-
-from . import rest
+from ..utils import rest
+from .forms import LoginForm
 
 
-bp = Blueprint('auth', __name__, url_prefix='/auth')
+mod_auth = Blueprint('auth', __name__, url_prefix='/auth')
 
 
-class LoginForm(FlaskForm):
-    email = StringField('Email', validators=[
-        InputRequired(),
-        Email(message='Not a valid email address.')
-    ])
-    password = PasswordField('Password', validators=[
-        InputRequired(),
-        Length(5, 64)
-    ])
-    submit = SubmitField('Sign In')
-
-
-@bp.route('/login', methods=['GET', 'POST'])
+@mod_auth.route('/login', methods=['GET', 'POST'])
 def login():
     form = LoginForm()
     if form.validate_on_submit():
@@ -45,14 +29,14 @@ def login():
         if error is None:
             session.clear()
             session['auth_token'] = auth_token
-            return redirect(url_for('index'))
+            return redirect(url_for('items.index'))
 
         flash(error)
 
     return render_template('auth/login.html', form=form)
 
 
-@bp.before_app_request
+@mod_auth.before_app_request
 def load_logged_in_user():
     auth_token = session.get('auth_token')
 
@@ -69,7 +53,7 @@ def load_logged_in_user():
             return redirect(url_for('auth.login'))
 
 
-@bp.route('/logout')
+@mod_auth.route('/logout')
 def logout():
     session.clear()
     return redirect(url_for('auth.login'))
@@ -80,8 +64,5 @@ def login_required(view):
     def wrapped_view(**kwargs):
         if g.user is None:
             return redirect(url_for('auth.login'))
-
         return view(**kwargs)
-
     return wrapped_view
-
