+++ front/app/search/forms.py
@@ -1,0 +1,11 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField, SubmitField, DateField
+from wtforms.validators import Optional
+from .. import utils
+
+
+class SearchForm(FlaskForm):
+    start_date = DateField('From', validators=[Optional()], default=utils.start_of_year, format="%Y-%m-%d")
+    end_date = DateField('Until', validators=[Optional()])
+    description = StringField('Description', validators=[Optional()])
+    submit = SubmitField('Search')
