+++ front/app/templates/base.html
@@ -17,7 +17,7 @@
   {% if g.user %}
     <nav class="navbar navbar-expand-sm navbar-light">
       <div class="container-xxl">
-        <a href="{{ url_for('index') }}" class="navbar-brand">fastBudsjett <img src="/static/img/piggy-bank-svgrepo-com.svg" width="28px" height="28px" style="border-width: 3px;"></img></a>
+        <a href="{{ url_for('items.index') }}" class="navbar-brand">fastBudsjett <img src="/static/img/piggy-bank-svgrepo-com.svg" width="28px" height="28px" style="border-width: 3px;"></img></a>
         <button
             class="navbar-toggler"
             type="button"
