+++ front/app/templates/items/remove.html
@@ -6,7 +6,7 @@
 <div class="container-xxl">
   <div class="row">
     <div class="col">
-      <a class="btn" href="{{ url_for('index') }}">
+      <a class="btn" href="{{ url_for('items.index') }}">
         <i data-feather="arrow-left-circle" width="28" height="28"></i>
       </a>
     </div>
