+++ front/app/templates/users/edit.html
@@ -11,7 +11,7 @@
     <div class="form-row justify-content-center my-2">
       <div class="mb-3 input-group">
         <span class="input-group-text">
-          <i class="bi bi-translate"></i>
+          <i class="bi bi-envelope"></i>
         </span>
         {{ form.email(class_="form-control", placeholder="e.g. bob@example.com") }}
       </div>
