+++ front/run.py
@@ -1,0 +1,2 @@
+from app import app
+app.run(host='0.0.0.0', port=5000, debug=True)
